feat(libc): real symlink/readlink/lstat/fstatat implementations
- symlink / readlink replace -1 stubs (6 busybox callers blocked).

- lstat / fstatat: new files. lstat needed for busybox 'ls -l' on

  symlinks; fstatat for newer glibc binaries.

- All via syscall3 / syscall4 (4-arg r10 ABI for fstatat's flags);

  negative return → errno = -ret.

Spec: docs/superpowers/specs/2026-09-05-symlink-support-design.md v5

§5.4.

+++ libc/include/unistd.h
@@ -71,7 +71,7 @@ int64_t execle(const char *path, const char *arg, ...);
 int unlink(const char *path);
 int link(const char *oldpath, const char *newpath);
 int symlink(const char *target, const char *linkpath);
-int readlink(const char *path, char *buf, size_t buf_size);
+ssize_t readlink(const char *path, char *buf, size_t buf_size);
 int mkdir(const char *path, int mode);
 int rmdir(const char *path);
 int rename(const char *oldpath, const char *newpath);
